Number of changes -- see notes below.
1. Convert the CONTAINS macro into CONTAINS_ABOVE and
CONTAINS_BELOW for different types of contain checks.

2. Eliminated all signed/unsigned comparison warnings

3. Maintain all canonicalized stack offsets as positive values

4. Cleaned up & added missing edits

5. Fixed API that grabs a system call's return value to return a long
type instead of int

+++ include/process.h
@@ -187,7 +187,7 @@ public:
    * @param retval output parameter into which the exit code is written
    * @return a return code describing the outcome
    */
-  ret_t getSyscallReturnValue(int &retval) const;
+  ret_t getSyscallReturnValue(long &retval) const;
 
   /**
    * Read 8 bytes of data from a virtual memory address.
